// Transitional (Editor Flow: fixing Short indices issue & more)

+++ app/src/androidTest/java/com/qubacy/moveanddraw/data/drawing/model/_test/util/DataDrawingGeneratorUtil.kt
@@ -5,7 +5,7 @@ import com.qubacy.moveanddraw.data.drawing.model.DataDrawing
 object DataDrawingGeneratorUtil {
     fun generateDataDrawingWithVerticesAndFaces(
         vertexArray: FloatArray,
-        faceArray: Array<Array<Triple<Short, Short?, Short?>>>
+        faceArray: Array<Array<Triple<Int, Int?, Int?>>>
     ): DataDrawing {
         return DataDrawing(
             vertexArray,
