+++ app/src/main/java/com/qubacy/moveanddraw/ui/application/activity/screen/editor/component/canvas/data/face/FaceSketch.kt
@@ -2,7 +2,7 @@ package com.qubacy.moveanddraw.ui.application.activity.screen.editor.component.c
 
 data class FaceSketch(
     val vertexArray: Array<Triple<Float, Float, Float>>,
-    val face: Array<Triple<Short, Short?, Short?>>
+    val face: Array<Triple<Int, Int?, Int?>>
 ) {
 
 }
